+++ migrations/20171212200650_add_admin_field_to_user.js
@@ -6,6 +6,9 @@ exports.up = function(knex, Promise) {
 };
 
 exports.down = function(knex, Promise) {
+    if (process.env.NODE_ENV == 'production') {
+        throw "Can't drop anything on production environment.";
+    }
     return knex.schema.table('users', function(t) {
         t.dropColumn('admin');
     });
